Bugfixes after merge commit - and added external API URI's

+++ PackagePlanner/Models/DeliveryData.cs
@@ -3,7 +3,7 @@
     public struct DeliveryData
     {
         public bool success;
-        public int price;
-        public int time;
+        public double price;
+        public double time;
     }
 }
